footer analytics added - gaps fixed
